test_case_enhancer/requirements.txt
# Test Case Enhancer Dependencies

# Browser automation
playwright>=1.40.0

# LLM Providers
google-generativeai>=0.3.0
openai>=1.0.0

# Utilities
pydantic>=2.0.0
